requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
google-generativeai
paddlepaddle
paddleocr
easyocr
pytesseract
torch
transformers
pillow
